--- src/main/java/echoServer/CRLF.java
@@ -1,7 +1,0 @@
-package echoServer;
-
-public class CRLF {
-
-    public static String CRLF = "\r\n";
-
-}
