--- src/main/java/echoServer/ClientReadable.java
@@ -1,8 +1,0 @@
-package echoServer;
-import java.io.IOException;
-
-public interface ClientReadable {
-
-    public String readLine() throws IOException;
-
-}
