--- src/main/java/echoServer/ClientReadableFactory.java
@@ -1,9 +1,0 @@
-package echoServer;
-import java.io.IOException;
-import java.net.Socket;
-
-public interface ClientReadableFactory {
-
-    public ClientReadable makeReader(Socket clientSocket) throws IOException;
-
-}
