--- src/main/java/echoServer/http/HTTPMethods.java
@@ -1,12 +1,0 @@
-package echoServer.http;
-
-public enum HTTPMethods {
-    OPTIONS,
-    GET,
-    HEAD,
-    POST,
-    PUT,
-    DELETE,
-    TRACE,
-    CONNECT;
-}
