--- src/main/java/echoServer/outputManagement/ClientWriteable.java
@@ -1,8 +1,0 @@
-package echoServer.outputManagement;
-import java.io.IOException;
-
-public interface ClientWriteable {
-
-    public void println(String message) throws IOException;
-
-}
