--- src/main/java/echoServer/outputManagement/ClientWriteableFactory.java
@@ -1,9 +1,0 @@
-package echoServer.outputManagement;
-import java.io.IOException;
-import java.net.Socket;
-
-public interface ClientWriteableFactory {
-
-    public ClientWriteable makePrinter(Socket clientSocket) throws IOException;
-
-}
