--- src/main/java/echoServer/outputManagement/ClientWriterFactory.java
@@ -1,10 +1,0 @@
-package echoServer.outputManagement;
-import java.io.IOException;
-import java.net.Socket;
-
-public class ClientWriterFactory implements ClientWriteableFactory {
-
-    public ClientWriteable makePrinter(Socket clientSocket) throws IOException {
-        return new PrintWriterWrapper(clientSocket);
-    }
-}
