--- src/main/java/echoServer/socketManagement/ServerSocketInterface.java
@@ -1,14 +1,0 @@
-package echoServer.socketManagement;
-
-import java.io.IOException;
-import java.net.Socket;
-
-public interface ServerSocketInterface {
-
-    Socket accept() throws IOException;
-
-    void bind(int port) throws IOException;
-
-    void close() throws IOException;
-
-}
