+++ src/main/java/httpServer/BufferedReaderWrapper.java
@@ -1,0 +1,17 @@
+package httpServer;
+import java.io.BufferedReader;
+import java.io.IOException;
+
+public class BufferedReaderWrapper implements ClientReadable {
+
+    public BufferedReader bufferedReader;
+
+    public BufferedReaderWrapper(BufferedReader bufferedReader) throws IOException {
+        this.bufferedReader = bufferedReader;
+    }
+
+    public String readLine() throws IOException {
+        return bufferedReader.readLine();
+    }
+
+}
