+++ src/main/java/httpServer/CRLF.java
@@ -1,0 +1,7 @@
+package httpServer;
+
+public class CRLF {
+
+    public static String CRLF = "\r\n";
+
+}
