+++ src/main/java/httpServer/ClientReadable.java
@@ -1,0 +1,8 @@
+package httpServer;
+import java.io.IOException;
+
+public interface ClientReadable {
+
+    public String readLine() throws IOException;
+
+}
