+++ src/main/java/httpServer/ClientReadableFactory.java
@@ -1,0 +1,9 @@
+package httpServer;
+import java.io.IOException;
+import java.net.Socket;
+
+public interface ClientReadableFactory {
+
+    public ClientReadable makeReader(Socket clientSocket) throws IOException;
+
+}
