+++ src/main/java/httpServer/Echoable.java
@@ -1,0 +1,11 @@
+package httpServer;
+import java.io.IOException;
+import java.net.Socket;
+
+public interface Echoable {
+    public Socket startServer() throws IOException;
+
+    public void readClientInput(Socket clientSocket) throws IOException;
+
+    public Socket keepListening() throws IOException;
+}
