+++ src/main/java/httpServer/http/HTTPMethods.java
@@ -1,0 +1,12 @@
+package httpServer.http;
+
+public enum HTTPMethods {
+    OPTIONS,
+    GET,
+    HEAD,
+    POST,
+    PUT,
+    DELETE,
+    TRACE,
+    CONNECT;
+}
