+++ src/main/java/httpServer/outputManagement/ClientWriteable.java
@@ -1,0 +1,8 @@
+package httpServer.outputManagement;
+import java.io.IOException;
+
+public interface ClientWriteable {
+
+    public void println(String message) throws IOException;
+
+}
