+++ src/main/java/httpServer/outputManagement/ClientWriteableFactory.java
@@ -1,0 +1,9 @@
+package httpServer.outputManagement;
+import java.io.IOException;
+import java.net.Socket;
+
+public interface ClientWriteableFactory {
+
+    public ClientWriteable makePrinter(Socket clientSocket) throws IOException;
+
+}
