+++ src/main/java/httpServer/outputManagement/ClientWriterFactory.java
@@ -1,0 +1,10 @@
+package httpServer.outputManagement;
+import java.io.IOException;
+import java.net.Socket;
+
+public class ClientWriterFactory implements ClientWriteableFactory {
+
+    public ClientWriteable makePrinter(Socket clientSocket) throws IOException {
+        return new PrintWriterWrapper(clientSocket);
+    }
+}
