+++ src/main/java/httpServer/socketManagement/ServerSocketInterface.java
@@ -1,0 +1,14 @@
+package httpServer.socketManagement;
+
+import java.io.IOException;
+import java.net.Socket;
+
+public interface ServerSocketInterface {
+
+    Socket accept() throws IOException;
+
+    void bind(int port) throws IOException;
+
+    void close() throws IOException;
+
+}
