--- src/test/java/echoServer/ClientWriterFactoryTest.java
@@ -1,23 +1,0 @@
-package echoServer;
-import echoServer.outputManagement.ClientWriteableFactory;
-import echoServer.outputManagement.ClientWriterFactory;
-import echoServer.outputManagement.ClientWriteable;
-import org.junit.jupiter.api.Test;
-
-import java.io.IOException;
-import java.net.ServerSocket;
-import java.net.Socket;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-public class ClientWriterFactoryTest {
-
-    @Test
-    void testMakePrinterReturnsAPrintWriter() throws IOException {
-        ServerSocket server = new ServerSocket(8080);
-        Socket socket = new Socket("localhost", 8080);
-        ClientWriteableFactory clientWriterFactory = new ClientWriterFactory();
-        ClientWriteable printer = clientWriterFactory.makePrinter(socket);
-        assertInstanceOf(ClientWriteable.class, printer);
-    }
-}
