+++ src/test/java/httpServer/EchoServerTest.java
@@ -1,0 +1,23 @@
+package httpServer;
+
+import httpServer.socketManagement.ServerSocketInterface;
+import httpServer.socketManagement.ServerSocketWrapper;
+import org.junit.jupiter.api.Test;
+import java.io.IOException;
+import java.net.Socket;
+import static org.junit.jupiter.api.Assertions.*;
+
+class EchoServerTest {
+
+    @Test
+    void testClientSocketClosesWhenCloseSocketIsCalled() throws IOException {
+        Socket clientSocket = new Socket();
+        ServerSocketInterface serverSocket = new ServerSocketWrapper();
+        EchoServer echoServer = new EchoServer();
+
+        echoServer.closeSockets(clientSocket, serverSocket);
+
+        assertTrue(clientSocket.isClosed());
+    }
+}
+
